+++ R/MODIS-package.R
@@ -18,7 +18,6 @@
 #' @importFrom graphics abline box grid locator title
 #' @importFrom mapedit drawFeatures selectFeatures
 #' @importFrom maps map.axes
-#' @importFrom maptools checkPolygonsHoles
 #' @importFrom methods as new slot
 #' @importFrom stats na.omit setNames smooth.spline
 #' @importFrom utils capture.output download.file installed.packages read.csv read.table vi write.csv write.table